backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
boto3==1.35.72
python-multipart==0.0.19
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
PyJWT==2.8.0
pillow==11.0.0
PyPDF2==3.0.1
PyMuPDF==1.24.13
reportlab==4.2.5
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
mangum==0.19.0
